feat(netxlite): use *Netx for NewParallelUDPResolver
The overall idea at this point is that we will have an interface
containing the methods we currently have inside *Netx and we will
sever the hard dependency between measurexlite and netxlite.

We will probably still need to import some contants from netxlite
but, for all effects and purposes, measurexlite would become
completely mockable (as it is now) with an explicit dependency
on the related behavior.

In turn, this refactoring would clarify which netxlite functionality
we use for measuring, which are legacy, and which we use for
communicating with the OONI backend. That change would definitely
make it obvious to modify netxlite and go in the overall direction
defined by https://github.com/ooni/probe/issues/2531.