refactor(measurexlite): depend on model.MeasuringNetwork (#1260)
With this diff, we detach measurexlite from netxlite. It was already
quite detached through functions used for testing. However, the changes
we implement are allow us to test measurexlite code by changing the
.Netx field of a *Trace, which means we can avoid using netxlite's
singleton for new code.

Another benefit of this diff is that we have clearly spelled out and
packaged into an interface the dependencies required to perform
measurements with measurexlite. Therefore, we can gracefully continue
separating the code used for measuring from the code for contacting the
backend, as detailed in https://github.com/ooni/probe/issues/2531.