cleanup: move measurex and tracex inside ./internal/legacy
This diff moves the measurex and tracex packages inside of the
./internal/legacy package. The reason why I am doing this is because
we don't plan on further developing these packages. We will just
fix bugs if they arise.

I am doing this change before looking into whether we could
more clearly figure out the dependencies these packages need
considering that we're trying to figure out and organize better
how we use netxlite throughout the codebase.

The overall objective of this work is to gently separate code
for measuring and code for communicating with the backend to
support https://github.com/ooni/probe/issues/2531.