feat(measurexlite): Trace now implements MeasuringNetwork
This diff is still yak shaving for https://github.com/ooni/probe/issues/2531
but produces some nice side effects. Now, we can either use netxlite.Netx
or measurexlite.Trace as a dependency for measuring code.

This change is absolutely in line with the original spirit of measurexlite
because it's now much simpler to write code for netxlite and then upgrade
to measurexlite.

I noticed this possibility yesterday while working on a much larger
diff, and now I am committing this smaller diff to start the working day.