feat(measurexlite): Trace now implements MeasuringNetwork (#1262)
This diff is still yak shaving for
https://github.com/ooni/probe/issues/2531 but produces some nice side
effects. Now, we can either use netxlite.Netx or measurexlite.Trace as a
dependency for measuring code.

The cost of this change is (1) switching some functions to use a
`model.DebugLogger`, which is not an issue, and (2) adding some optional
wrappers to other functions. I made these functions to ignore the
wrappers and documented this limitation, because measurexlite code does
not need this functionality. It is a bit debateable whether this is the
best way to implement this change, but I am doing this because I would
like to see these wrappers gone when we stop using legacy/netx in the
codebase, so I think it's fine to ignore them, to better signal they're
deprecated stuff.

This change is absolutely in line with the original spirit of
measurexlite because it's now much simpler to write code for netxlite
and then upgrade to measurexlite.

I noticed this possibility yesterday while working on a much larger
diff, and now I am committing this smaller diff to start the working
day.