refactor: rename geolocate enginelocate
I am soon going to start addressing the underlying issue described
by https://github.com/ooni/probe/issues/2531.

But, before doing that, I have noticed that the packets I need to
edit to this end are the following:

```
./internal/engine
./internal/geolocate
./internal/sessionresolver
```

Now, these three packages work in unison to provide an engine.Session
and they should sort together. However, I would like also to avoid
nesting because I think all of them servers an ~independent purpose
and geolocate and sessionresolver are possibly building blocks to
refactor or reimplement the engine.

For this reason, I have chosen to rename them such that it is clear
they are the engine and supporting packages.

This diff addresses the first half of the change by renaming the
geolocate package.

While there acknowledge that the script to rename packages was broken
and decide to ditch it for good rather than entering into the quest
of fixing it. I would probably have spent lots of time trying in doing
that, and my time is better spent otherwise. (I remember the official
package renaming tool did not support go modules, but probably it
does and I should look into giving it another spin.)