refactor: rename sessionresolver engineresolver (#1264)
I am soon going to start addressing the underlying issue described by
https://github.com/ooni/probe/issues/2531.

But, before doing that, I have noticed that the packages I need to edit
to this end are the following:

```
./internal/engine
./internal/geolocate
./internal/sessionresolver
```

Now, these three packages work in unison to provide an `engine.Session`
and they should *sort* together. However, I would like also to avoid
nesting because I think each of them serves an ~independent purpose, and
geolocate and sessionresolver are possibly building blocks to refactor
or reimplement the engine.

For this reason, I have chosen to rename them such that it is clear they
are the engine and supporting packages.

This diff addresses the second half of the change by renaming the
sessionresolver package.