feat(netxlite): add flexible HTTP transport factory
Now that we've clearly labeled and package technical debt, we can
copy existing technical-debt-ridden code and modify it to obtain a
flexible factory for creating HTTP transports.

In particular, this factory uses sensible defaults for measuring
and there are options that you can pass it to customize details
such as the backend proxy as well as other underlying behavior that
we previously unconditionally configured.

This factory will be the starting point for having custom network
functions for the engine in line with https://github.com/ooni/probe/issues/2531.