feat(netxlite): add flexible HTTP transport factory (#1270)
Now that we've clearly labeled and packaged technical debt, we can copy
existing technical-debt-ridden code and modify it to obtain a flexible
factory for creating HTTP transports, `NewHTTPTransportWithOptions`.

In particular, this factory uses sensible defaults for measuring and
there are options that you can pass it to customize details such as the
backend proxy that we previously unconditionally configured.

More in detail, this is a side-by-side comparison between new code's
defaults and old code:

| Setting Name | httpquirks.go (old code) | httpfactory.go (new code) |
| ------------------- | ------------------------ |
-------------------------- |
| .Proxy | nil | nil |
| .MaxConnsPerHost | 1 | ooni/oohttp's default |
| .DisableCompression | true | true |
| .ForceAttemptHTTP2 | true | true |

So, basically, the biggest change is that we've removed the limitation
of the max number of connections per host being set to 1. In any case,
the new code allows to configure each of these fields.

This factory will be the starting point for having custom network
functions for the engine in line with
https://github.com/ooni/probe/issues/2531.

While there, acknowledge that `NewHTTPClient` contains technical debt
because it does unconditionally disable cookies, to document that and
move it inside of the proper file (`httpquirks.go`).