feat(netxlite): add (*Netx).ListenTCP
This diff adds the ListenTCP function to the *Netx struct.

With this addition, based on https://github.com/ooni/probe-cli/pull/1278, we
can remove some techdebt inside the testingx package.

This is yak shaving while trying to add support for testing socks5
in the context of https://github.com/ooni/probe/issues/2531.