feat: support for testing socks5 clients
This diff imports a fork of github.com/armon/go-socks5 that has been
adapted to use netem and simplified to suit our testing needs.

With this functionality in tree, we can start thinking about writing
better netem based tests for the ooniprobe bootstrap.

The overall idea is to be well positioned to improve the bootstrap and
introduce dynamic support for beacons.

Reference issue: https://github.com/ooni/probe/issues/2531