refactor(model/netx.go): TLSHandhaker now returns a TLSConn
I am making progress with https://github.com/ooni/probe/issues/2531
and I want to reactor model/netx.go such that the TLSHandshaker returns
a model.TLSConn rather than a net.Conn.

Returning a net.Conn and documenting it is a model.TLSConn is bad
compared to returning a model.TLSConn directly.

Note that we cannot apply the same transformation to netxlite's
TLSDialer.DialTLSContext because such a method must be assignable
to net/http and github.com/ooni/oohttp's Transport function also
called DialTLSContext.

The fact that we need code to be assignable to the Transport
function is what historically led the TLSHandshaker to return
a net.Conn as well. But it was quite clear from the get go
that this choice led to some quirks (and, in fact, this behavior
was explicitly documented as such). Not anymore.