feat(enginenetx): support HTTP and HTTPS proxies
This diff completes the work we have been doing for a few days
now and provides HTTP and HTTPS proxy support, in addition to
SOCKS5 support, for the engine-specific network.

We did this work in the context of https://github.com/ooni/probe/issues/2531
and https://github.com/ooni/probe/issues/1955.