feat(enginenetx): support HTTP and HTTPS proxies (#1282)
This diff completes the work we have been doing for a few days now and
provides HTTP and HTTPS proxy support, in addition to SOCKS5 support,
for the engine-specific network.

We did this work in the context of
https://github.com/ooni/probe/issues/2531 and
https://github.com/ooni/probe/issues/1955.

BTW, the fact that tests used `measurexlite` and tracing is very nice
here. It means the idea to write `measurexlite` based on context and
tracing was good and could be used beyond its original design goals.