feat(enginenetx): make sure HTTPSDialer closes all connections (#1285)
This uses the code introduced in the previous commit, i.e.,
https://github.com/ooni/probe-cli/commit/ac13e53521a4317d35239e8c7bf969f738816cf4.

I think it's good to have additional confidence that the `HTTPSDialer`
does not leak connections. (Ideally, we would like to verify that we're
not leaking connections everywhere, but for now just doing it for new
code would do okay.)

While there, repair a test that was flaky because of nondeterministic
channel selection.

Part of https://github.com/ooni/probe/issues/2531.