fix(enginenetx): pass context to tactics callbacks
This diff modifies the tactics callbacks to take in input a context.

We need the context to know whether the operation failed in itself or
was canceled externally through the context.

With this information, we can store better statistics about which
tactics are working and which tactics are not working.

Part of https://github.com/ooni/probe/issues/2531