refactor(enginenetx): introduce stats
This diff refactors the code to introduce a stats interface and
to make the tactic a struct ~without methods attached.

We transform the tactic into a struct because we're planning on storing
the tactics on disk and loading them.

We need stats to track what is working and choose which is the
best tactic that we should employ.

In light of these two needs, it makes sense to transfer the reporting
methods that previously were part of the tactic interface to stats.

Part of https://github.com/ooni/probe/issues/2531