fix(enginenetx): refine the happy-eyeballs algorithm
We want to pack attempts in parallel, which we also did before
when the interval between attempts was linear.

We need to take into account possible congestion, so we should push
back exponentially, even though the common case for us is probably
censorship (we it is better to do the right thing anyway).

So, let's scale exponentially until we reach 30s. After that, it's
fine to keep attempts evenly spaces, because that is quite a huge
interval if we're reasoning in internet time.

Also, change the base value used for TLS handshaking to be 900ms
rather than 300ms, because a TLS handshake is ~3 round trips.

Part of https://github.com/ooni/probe/issues/2531