refactor(enginenetx): make static/loadable policy easier to use
This diff makes the policy previously known as loadable and now know
as the static policy easier to use, by having a constructor that reads
from a key-value store and by passing it a fallback policy to use.

With this design, it should be possible to have code that uses the
static policy if applied, falling back to whatever policy we are
otherwise constructing into the NewNetwork constructor.

In a subsequent commit, I will hook this code into NewNetwork, so
that we can override the policy by changing the filesystem.

Part of https://github.com/ooni/probe/issues/2531.