refactor(enginenetx): make LookupTactics async (#1300)
Rather than waiting for LookupTactics to complete, make it async and let
it stream the tactics back to the caller. This design allows us to start
connecting while DNS lookups are still in progress when we have
configured beacons for specific hosts. In turn, this means we could
perform more operations in the same unit of time, by overalapping some
DNS lookups and TCP+TLS dials. Additionally, the new design would also
work quite well with a DNS resolver that awaits for additional responses
after the first one and returns all of them as tactics.

While there, recognize that the HTTPSDialer code and the code in the
related structs was a bit more complex than it should be. We don't need
to explicitly honor the context when moving data between goroutines as
long as the writer goroutines write until completion and then close the
channel, and as long as reader goroutines read until either the channel
is closed (when there's a single writer) or all the possible writers
have completed (otherwise). Networking code and networking-like code is
the only code that MAY block and for which we really need a context.

With the new simplified design, all the goroutines will join before
`DialTLSContext` returns, hence we don't need anymore a `sync.WaitGroup`
to make sure we're not leaking any goroutine in this code.

Part of https://github.com/ooni/probe/issues/2531