fix(enginenetx): improve happy eyeballs algorithm
The current algorithm is broken in that, after doubling the delay for
a while (a) we could overflow if index is too big and (b) we end up
scheduling all attempts 30 seconds after the beginning.

Rather, we should switch to increasingly the delay linearly after we
have reached a given cutoff and we should not use an algorithm that
depends on the number be small to avoid overflow.

Part of https://github.com/ooni/probe/issues/2531