feat(enginenetx): prune old entries from stats
If an entry is older than one week ago, we remove it from the stats.

This helps ensuring the stats cache does not grow too large.

Additionally, this helps ensuring we are not going to try tactics that
worked long time ago and it may be pointless trying now.

Part of https://github.com/ooni/probe/issues/2531