refactor(enginenetx): store address and port separately
This diff modifies the `HTTPSDialerTactic` struct to store the
address and port separately rather than storing the endpoint as
the concatenation of the address and the port.

By storing these two fields separately, we're more flexible as we
will soon see with subsequent diffs.

Reference issue: https://github.com/ooni/probe/issues/2531