refactor(enginenetx): organize by domain and endpoint
There's no point in using "www.example.com" as the key with policies
for mixed ports when actually we should be using "www.example.com:443"
such that we only group policies for the same port.

This needs to happen both for the data format of the stats stored
to disk and for the user-provided static policy format.

Part of https://github.com/ooni/probe/issues/2531