refactor(enginenetx): make static-policy API private
Now that we've more or less reached the point where we wanted to be with
https://github.com/ooni/probe/issues/2531, let's spend some time to
refactor the implementation, now that we know very well how to works,
such that modifying it in the future would be easier.

The first order of business here seems to hide implementation details
and get rid of too many HTTPSDialer prefixes which only cause confusion
when looking at the available structs.

Here, in particular, we deal with the static-policy API. To this end,
we also need to more some tests around because they need internals.

Also, while there, add a top-level description to some files, so that
newcomers have an idea of what each file is about.