feat(enginenetx): support getting stats on a domain endpoint
We need this functionality to ask the stats questions like "what are
all the available tactics for www.example.com:443 along with their
stats?". In turn, this information is needed to implement a new policy
that uses existing stats to choose what to do.

Part of https://github.com/ooni/probe/issues/2531