feat(enginenetx): add policy based on stats
This policy first checks the stats and sorts the entries for which
we have the highest success rate. Then, we use an underlying policy
to produce suitable tactics for the HTTPS dialer.

While there, rename statstracker{,_test}.go to statsmanager{,_test}.go
because the type inside it is called "state manager". The state
tracker is an abstract concept only used by the https dialer code.

Part of https://github.com/ooni/probe/issues/2531