fix(enginenetx): use a simple policy with proxy
This diff modifies how we construct a `*Network` to use a very
simple DNS-only policy when using a proxy.

We don't need to try anything fancy when we're using a proxy because we
assume the proxy knows how to do circumvention and we don't to get in
the proxy way.

While there, recognize that the static policy was named very
similarly to the stats policy, so rename the former user policy,
such that there is more lexicographical difference.

Part of https://github.com/ooni/probe/issues/2531