fix(enginenetx): periodically trim statistics (#1317)
If we're running the enginenetx code inside an environment that sprays
~random IP addresses during DNS lookup, we're going to end up using
quite a large amount of statistics. While this amount of statistics
could potentially be great in general, we need to be mindful of memory
and disk occupation as well as of the overhead of serializing and
deserializing.

To address these concerns, add the arbitrary rule that we don't want
more than ten domain-endpoint entries and implement code to zap old
and/or always-failing entries. The code will run when loading the
entries, when serializing them, and periodically while the `*Network` is
up and running.

See https://github.com/ooni/probe/issues/2531