refactor: move scrubbing logger to logx
This diff is yak shaving for a subsequent diff that attempts to
mitigate potential IP addresses leaks caused by using happy eyeballs
more aggressively in the codebase. The general idea is that we
previously had a netxlite implementation that gave IPv4 preference
over IPv6, meaning that we would end up using IPv4 in most cases
and very rarely IPv6. As part of my work to make beacons possible,
I have introduced happy eyeballs, which means we may sometimes
use IPv4 instead of IPv6. This fact makes it more likely that we
include the IPv6 address of a probe when we know its IPv4 address
or the other way around into measurements. To mitigate this possible
issue before it actually is possible (note that I have not yet
changed how we discover the probe IP), I am going to proactively
modify the serialization of HTTP bodies and headers to scrub. However,
to do this, I need to decouple the scrubber package from model such
that internal/model/archival.go can use the scrubber package.

Part of https://github.com/ooni/probe/issues/2531