feat(model): introduce ArchivalBinaryData
There are cases where we know we have binary data in input and we
want the output to be binary data using the dictionary encoding like
`{"format":"base64","data":"..."}`.

Such cases are, for example, DNS messages bodies.

There is no need for us to pass through the MaybeArchivalBinaryData
in such cases. Additionally, this makes MaybeArchivalBinaryData a bit
more complex than it probably needs to be.

What's more, ArchivalBinaryData is for when you do not require to
scrub the data. I want to introduce new data types that automatically
perform scrubbing when they're used.

But this puts even more pressure on MaybeArchivalBinaryData, and
hence this commit, to start differentiating between:

1. always binary data vs maybe binary data

2. no scrubbing requires vs scrubbing required

The rationale for doing this set of changes has been explained in
https://github.com/ooni/probe-cli/pull/1319.

The reference issue is https://github.com/ooni/probe/issues/2531.

For now, this commit just adds the new type and tests for it
without using the type, which we'll do later once we have added
better marshal/unmarshal testing for the interested types.