refactor(model): simplify using ArchivalBinaryData
It did not originally occur to me, but now it's clear that we can
avoid using a struct to wrap the data type.

It just suffices to use the new type.

I think this is better in terms of writing code because the only
two things we need to do are:

1. make sure we have serialization and unserialization tests;

2. use the correct data type in the struct.

For all intents and purposes the ArchivalBinaryData is just a special
kind of []byte attached to custom marshal/unmarshal rules.

Part of https://github.com/ooni/probe/issues/2531