feat: add ArchivalMaybeBinaryString type
This type is ~equivalent to ArchivalMaybeBinaryData but designed to hold
a string, do less type conversions, and be easier to use.

The intent is to replace the ArchivalMaybeBinaryData type with this type.

For now, let us introduce the new type and its tests.

Part of https://github.com/ooni/probe/issues/2531