cleanup: common funcs for setting headers (#1332)
I was trying to make sure the code was using a new definition of headers
based on ArchivalMaybeBinaryString, but I stumbled upon a roadblock
where there were too many code paths settings headers.

I did not feel comfortable changing multiple code paths, so here is
instead a diff that unifies setting headers.

The slightly tricky part of this diff has been the requirement that we
preserve headers' capitalization for hhfm. The catch seems to be that we
should not use http.Header for setting the headers in hhfm, but rather
pass through map[string][]string, which is not attached case
normalization setters. I think we're fine in this department because we
already had test cases and I added a bunch more test cases.

After this diff is merged, I can resume with my plan to make headers use
ArchivalMaybeBinaryString, which, in turn, is functional to
automatically scrub headers, which is something we would like to happen
in light of the introduction of more happy eyeballs code due to
https://github.com/ooni/probe/issues/2531.