cleanup: headers map repr now uses ArchivalMaybeBinaryString
This diff modifies the headers-as-a-map representation such that
we use ArchivalMaybeBinaryString instead of ArchivalMaybeBinaryData.

The overall idea is to migrate all users of the latter to the
former, such that we can remove the latter and implement extra
scrubbing for all ArchivalMaybeBinaryString users.

Part of https://github.com/ooni/probe/issues/2531