refactor: use ArchivalMaybeBinaryString for headers list
This commit changes the list representation of HTTP headers to
use the ArchivalMaybeBinaryString type.

Now that we have migrated most representations of HTTP headers to
use this new type, we can finally implement scrubbing.

As a reminder, we still need to migrate the ./legacy/measurex
implementation, however that one is a bit annoying because there
are no tests at all for measurex 🤦🤦🤦🤦🤦🤦🤦.

That said, given that ./legacy/measurex is deprecated and it
is only used by the tor experiment, I think leaving it untouched
would probably be the right thing to do here.

Part of https://github.com/ooni/probe/issues/2531