feat(archival): unconditionally scrub HTTP headers and bodies (#1335)
This commit modifies ArchivalMaybeBinaryData to unconditionally scrub IP
addresses and endpoints. It also renames the struct
ArchivalScrubbedMaybeBinaryData to underline that it's scrubbing now.

We're using ArchivalScrubbedMaybeBinaryData to represent HTTP headers
and bodies collected by OONI measurements.

With this commit merged, I am now much less concerned about the
potential unintended effects of aggressively using happy eyeballs, which
we started introducing as part of
https://github.com/ooni/probe/issues/2531.

In other words, the rest of the refactoring of the OONI bootstrap could
proceed a bit faster.