fix: make bootstrap more robust (and slower)
This diff ensures that the boostrap is more robust in the scenario
described by https://github.com/ooni/probe/issues/2544.

Because we're close to a release, I did not want to write a heavy
refactor of the boostrap, and specifically of the iplookup part, but
I documented what is needed at https://github.com/ooni/probe/issues/2551.

This diff has been tested in a problematic network under two
distinct conditions:

1. cold bootstrap (i.e., w/o prior knowledge), where the correct
fix is to give the system resolver intermediate priority;

2. hot bootstrap with fscked up config, where the correct fix
is to use a large coarse grained timeout.

The fix that causes a canceled context to prevent marking a resolver
as failed helps in both scenarios.

As you can see, I removed a "reliability fix", which actually was
more of an optimization. This removal means that the probe bootstrap
is now slower than it could be, hence my choice of documenting
what I'd do had I had more time to work on this topic.